Split userContext into seperate classes
