+++ Nonogram/Database/UserContext.cs
@@ -7,53 +7,54 @@ using System.Text.Json;
 using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 using Nonogram.Models;
+using Nonogram.Interfaces;
 
 namespace Nonogram.Database
 {
-    interface IUser
-    {
-        public void SaveToUser(User user, string filePath);
-        public void SaveNewUser(User user, string filePath);
-        public List<User> GetUsers(string filePath);
-    }
+    //TODO: REMOVE CLASS
+    //interface IUser
+    //{
+    //    public void SaveToUser(User user, string filePath);
+    //    public void SaveNewUser(User user, string filePath);
+    //    public List<User> GetUsers(string filePath);
+    //}
 
+    //public class JsonUserDatabase : IUser
+    //{
+    //    readonly JsonSerializerOptions options = new JsonSerializerOptions { WriteIndented = true };
 
-    public class JsonUserDatabase : IUser
-    {
-        readonly JsonSerializerOptions options = new JsonSerializerOptions { WriteIndented = true };
+    //    public List<User> GetUsers(string filePath)
+    //    {
+    //        if (!File.Exists(filePath))
+    //            return new List<User>();
 
-        public List<User> GetUsers(string filePath)
-        {
-            if (!File.Exists(filePath))
-                return new List<User>();
+    //        string json = File.ReadAllText(filePath);
+    //        if (string.IsNullOrWhiteSpace(json))
+    //            return new List<User>();
 
-            string json = File.ReadAllText(filePath);
-            if (string.IsNullOrWhiteSpace(json))
-                return new List<User>();
+    //        List<User> tmp = JsonSerializer.Deserialize<List<User>>(json, options);
+    //        return tmp;
+    //    }
 
-            List<User> tmp = JsonSerializer.Deserialize<List<User>>(json, options);
-            return tmp;
-        }
+    //    public void SaveToUser(User user, string filePath)
+    //    {
+    //        List<User> users = GetUsers(filePath);
+    //        int indx = users.FindIndex(u => u.Name == user.Name);
 
-        public void SaveToUser(User user, string filePath)
-        {
-            List<User> users = GetUsers(filePath);
-            int indx = users.FindIndex(u => u.Name == user.Name);
+    //        if (indx != -1)
+    //            users[indx] = user;
 
-            if (indx != -1)
-                users[indx] = user;
+    //        string json = JsonSerializer.Serialize(users, options);
+    //        File.WriteAllText(filePath, json);
+    //    }
 
-            string json = JsonSerializer.Serialize(users, options);
-            File.WriteAllText(filePath, json);
-        }
+    //    public void SaveNewUser(User user, string filePath)
+    //    {
+    //        List<User> users = GetUsers(filePath);
+    //        users.Add(user);
 
-        public void SaveNewUser(User user, string filePath)
-        {
-            List<User> users = GetUsers(filePath);
-            users.Add(user);
-
-            string json = JsonSerializer.Serialize(users, options);
-            File.WriteAllText(filePath, json);
-        }
-    }
+    //        string json = JsonSerializer.Serialize(users, options);
+    //        File.WriteAllText(filePath, json);
+    //    }
+    //}
 }
